requirements.txt
python-dotenv~=0.20.0
praw~=7.6.0
requests==2.27.1
beautifulsoup4~=4.11.1
pyshorteners~=1.0.1